Fixes #559 for guests but also for other kinds of users (only applies to images displayed full screen)
